Update logic to fetch md file from GH directly and remove mdl json and script
